更新文件系统例程，修改sd卡驱动

+++ edriver/mmc_sd.h
@@ -16,22 +16,22 @@ public:
         this->cs  = cs;
         this->spi = spi;
     }
-    int begin(uint8_t dev_num);
-    uint8_t init();
-    int get_CID(u8 *cid_data);
-    int get_CSD(u8 *csd_data);
-    u32 get_capacity(void);
-    u8 read_single_block(u32 sector, u8 *buffer);
-    u8 write_single_block(u32 sector, const  u8 *data);
-    u8 read_multi_block(u32 sector, u8 *buffer, u8 count);
-    u8 write_multi_block(u32 sector,  const u8 *data, u8 count);
-    u8 read_bytes(unsigned long address, unsigned char *buf, unsigned int offset, unsigned int bytes);
+    int         begin(uint8_t dev_num);
+    uint8_t     init();
+    int         get_CID(u8 *cid_data);
+    int         get_CSD(u8 *csd_data);
+    uint64_t    get_capacity(void);
+    u8          read_single_block(u32 sector, u8 *buffer);
+    u8          write_single_block(u32 sector, const  u8 *data);
+    u8          read_multi_block(u32 sector, u8 *buffer, u8 count);
+    u8          write_multi_block(u32 sector,  const u8 *data, u8 count);
+    u8          read_bytes(unsigned long address, unsigned char *buf, unsigned int offset, unsigned int bytes);
 
 private:
-    uint8_t _wait(void);
-    uint8_t _send_command(u8 cmd, u32 arg, u8 crc);
-    uint8_t _send_command_no_deassert(u8 cmd, u32 arg, u8 crc);
-    int _receive_data(u8 *data, u16 len, u8 release);
+    uint8_t     _wait(void);
+    uint8_t     _send_command(u8 cmd, u32 arg, u8 crc);
+    uint8_t     _send_command_no_deassert(u8 cmd, u32 arg, u8 crc);
+    int         _receive_data(u8 *data, u16 len, u8 release);
 
 public:
     u8  SD_Type; //SD¿¨µÄÀàĞÍ
